changed directory for test_scoring.py file
